refactor: i moved/deleted some classes and functions to their respective directories to make it easy to find and navigate.
