+++ app/src/main/java/dev/iamwami/app/quotinews/data/local/dao/NewsDao.kt
@@ -1,9 +1,9 @@
-package dev.iamwami.app.quotinews.db.dao
+package dev.iamwami.app.quotinews.data.local.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
-import dev.iamwami.app.quotinews.db.entity.NewsTable
+import dev.iamwami.app.quotinews.data.local.entity.NewsTable
 import dev.iamwami.app.quotinews.model.News
 import kotlinx.coroutines.flow.Flow
