+++ app/src/main/java/dev/iamwami/app/quotinews/data/repository/NewsRepository.kt
@@ -1,6 +1,6 @@
 package dev.iamwami.app.quotinews.data.repository
 
-import dev.iamwami.app.quotinews.db.entity.NewsTable
+import dev.iamwami.app.quotinews.data.local.entity.NewsTable
 import kotlinx.coroutines.flow.Flow
 
 interface NewsRepository {
