+++ app/src/main/java/dev/iamwami/app/quotinews/di/DaoModule.kt
@@ -4,8 +4,8 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import dev.iamwami.app.quotinews.db.AppDatabase
-import dev.iamwami.app.quotinews.db.dao.NewsDao
+import dev.iamwami.app.quotinews.data.local.db.AppDatabase
+import dev.iamwami.app.quotinews.data.local.dao.NewsDao
 import javax.inject.Singleton
 
 @Module
