+++ app/src/main/java/dev/iamwami/app/quotinews/di/DatabaseModule.kt
@@ -7,7 +7,7 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import dev.iamwami.app.quotinews.db.AppDatabase
+import dev.iamwami.app.quotinews.data.local.db.AppDatabase
 import javax.inject.Singleton
 
 @Module
