+++ app/src/main/java/dev/iamwami/app/quotinews/navigation/QuotiNewsNavGraph.kt
@@ -10,8 +10,8 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import dev.iamwami.app.quotinews.ui.home.HomeRoute
-import dev.iamwami.app.quotinews.ui.home.HomeViewModel
+import dev.iamwami.app.quotinews.ui.screens.home.HomeRoute
+import dev.iamwami.app.quotinews.ui.screens.home.HomeViewModel
 import dev.iamwami.app.quotinews.ui.screens.bookmark.BookmarkRoute
 import dev.iamwami.app.quotinews.ui.screens.splashscreens.NavigateToSplashScreen
