+++ app/src/main/java/dev/iamwami/app/quotinews/ui/screens/components/RelatedNews.kt
@@ -21,11 +21,11 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import dev.iamwami.app.quotinews.db.entity.NewsTable
+import dev.iamwami.app.quotinews.data.local.entity.NewsTable
 import dev.iamwami.app.quotinews.ui.components.LikeBtn
 import dev.iamwami.app.quotinews.ui.components.NewsImage
 import dev.iamwami.app.quotinews.ui.components.PostDivider
-import dev.iamwami.app.quotinews.ui.utils.SampleNewsApiDataProvider
+import dev.iamwami.app.quotinews.util.SampleNewsApiDataProvider
 
 @Composable
 fun RelatedNews(
