+++ app/src/main/java/dev/iamwami/app/quotinews/util/SampleData.kt
@@ -1,4 +1,4 @@
-package dev.iamwami.app.quotinews.ui.utils
+package dev.iamwami.app.quotinews.util
 
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import dev.iamwami.app.quotinews.model.Articles
